+++ patientMatcher/match/handler.py
@@ -167,7 +167,6 @@ def external_matcher(database, patient, node=None):
 
     LOG.info("Matching patient against {} node(s)..".format(len(connected_nodes)))
     for node in connected_nodes:
-        server_name = node["_id"]
         node_url = node["matching_url"]
         token = node["auth_token"]
         request_content_type = node["accepted_content"]
@@ -177,7 +176,6 @@ def external_matcher(database, patient, node=None):
             "Accept": "application/vnd.ga4gh.matchmaker.v1.0+json",
             "X-Auth-Token": token,
         }
-        LOG.info('sending HTTP request to server: "{}"'.format(server_name))
         # send request and get response from server
         json_response = None
         server_return = None