+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Fixed
 - Broken Link to HPO definitions (`phenotype_annotation.tab`)
 - Security hotspots flagged by SonarCloud
+- Decreased vulnerability by removing some server log messages
 
 ## [4.1] - 2022-06-27
 ### Added